requirements.txt
# ======================
# Customer Management System
# Requirements Specification
# ======================

# === CORE & FRAMEWORK ===
# Flask Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# FastAPI (اگر نیاز دارید)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# === UI & WEB ===
# Flask Extensions
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Limiter==3.3.3
Flask-Caching==2.0.2
eventlet==0.33.3  # برای SocketIO

# === DEPLOYMENT SYSTEM ===
# Docker Management
docker==6.1.3

# SSH & Remote Execution
paramiko==3.3.1
fabric==2.7.1

# Configuration & Templating
PyYAML==6.0
python-dotenv==1.0.0
jinja2==3.1.2
python-crontab==2.7.1

# Task Queue (اگر نیاز به async tasks دارید)
celery==5.3.1
redis==4.6.0  # برای Celery backend

# === DATABASE & ORM ===
# SQLAlchemy (اگر از دیتابیس استفاده می‌کنید)
SQLAlchemy==2.0.23
alembic==1.12.1

# Database Drivers
psycopg2-binary==2.9.9  # برای PostgreSQL

# === SECURITY & AUTH ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# === UTILITIES ===
# HTTP Client
requests==2.31.0
httpx==0.25.2

# System Monitoring
psutil==5.9.6

# Data Processing (اگر نیاز دارید)
pandas==2.1.3
numpy==1.26.2

# Date/Time
pytz==2023.3.post1

# === TESTING ===
# Pytest Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-html==4.1.1
pytest-xdist==3.5.0

# API Testing
requests==2.31.0

# UI Testing
selenium==4.15.2
playwright==1.40.0
pytest-playwright==0.4.3

# Load Testing
locust==2.20.1

# Test Reporting
allure-pytest==2.13.2
pytest-check==2.2.2

# === OPTIONAL (بر اساس نیاز) ===
# برای document generation
# markdown==3.5.1

# برای گزارش‌گیری
# matplotlib==3.8.2

# برای export به Excel
# openpyxl==3.1.2

# برای logging پیشرفته
# structlog==23.2.0
# sentry-sdk==1.38.0